requirements.txt
paho-mqtt==2.1.0
python-dotenv==1.2.1
